// clients/admin-ui/src/types/api/models/TCFRestrictionType.ts
// This file is auto-generated by @hey-api/openapi-ts

/**
 * TCFRestrictionType
 *
 * Enum for TCF restriction types
 */
export enum TCFRestrictionType {
  PURPOSE_RESTRICTION = "purpose_restriction",
  REQUIRE_CONSENT = "require_consent",
  REQUIRE_LEGITIMATE_INTEREST = "require_legitimate_interest",
}