// clients/admin-ui/src/types/api/models/TCFVendorRestriction.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * Enum for TCF vendor restriction types
 */
export enum TCFVendorRestriction {
  RESTRICT_ALL_VENDORS = 'restrict_all_vendors',
  ALLOW_SPECIFIC_VENDORS = 'allow_specific_vendors',
  RESTRICT_SPECIFIC_VENDORS = 'restrict_specific_vendors',
}